+++ src/cli/notes.rs
@@ -1,9 +1,4 @@
-use std::{
-    collections::{HashMap, HashSet},
-    fs::File,
-    io::{Read, Write},
-    path::PathBuf,
-};
+use std::collections::{HashMap, HashSet};
 
 use clap::ValueEnum;
 use comfy_table::{presets, Attribute, Cell, ContentArrangement, Table};
@@ -15,10 +10,9 @@ use miden_client::{
 use miden_objects::{
     accounts::AccountId,
     crypto::rand::FeltRng,
-    notes::{NoteId, NoteInputs, NoteMetadata},
+    notes::{NoteInputs, NoteMetadata},
     Digest,
 };
-use miden_tx::utils::{Deserializable, Serializable};
 
 use super::{Client, Parser};
 use crate::cli::{
@@ -63,31 +57,6 @@ pub enum Notes {
         inputs: bool,
     },
 
-    /// Export note data to a binary file.
-    #[clap(short_flag = 'e')]
-    Export {
-        /// Note ID of the note to show. We only allow to export a note that has been created using
-        /// this client
-        #[clap()]
-        id: String,
-
-        /// Path to the file that will contain the note data. If not provided, the filename will be the input note ID
-        #[clap()]
-        filename: Option<PathBuf>,
-    },
-
-    /// Import note data from a binary file
-    #[clap(short_flag = 'i')]
-    Import {
-        /// Path to the file that contains the input note data
-        #[clap()]
-        filename: PathBuf,
-
-        /// Skip verification of note's existence in the chain
-        #[clap(short, long, default_value = "false")]
-        no_verify: bool,
-    },
-
     /// List consumable notes
     #[clap(short_flag = 'c')]
     ListConsumable {
@@ -97,10 +66,16 @@ pub enum Notes {
     },
 }
 
+impl Default for Notes {
+    fn default() -> Self {
+        Notes::List { filter: None }
+    }
+}
+
 impl Notes {
     pub async fn execute<N: NodeRpcClient, R: FeltRng, S: Store>(
         &self,
-        mut client: Client<N, R, S>,
+        client: Client<N, R, S>,
     ) -> Result<(), String> {
         match self {
             Notes::List { filter } => {
@@ -115,14 +90,6 @@ impl Notes {
             },
             Notes::Show { id, script, vault, inputs } => {
                 show_note(client, id.to_owned(), *script, *vault, *inputs)?;
-            },
-            Notes::Export { id, filename } => {
-                export_note(&client, id, filename.clone())?;
-                println!("Succesfully exported note {}", id);
-            },
-            Notes::Import { filename, no_verify } => {
-                let note_id = import_note(&mut client, filename.clone(), !(*no_verify)).await?;
-                println!("Succesfully imported note {}", note_id.inner());
             },
             Notes::ListConsumable { account_id } => {
                 list_consumable_notes(client, account_id)?;
@@ -161,65 +128,6 @@ fn list_notes<N: NodeRpcClient, R: FeltRng, S: Store>(
 
     print_notes_summary(zipped_notes)?;
     Ok(())
-}
-
-// EXPORT NOTE
-// ================================================================================================
-pub fn export_note<N: NodeRpcClient, R: FeltRng, S: Store>(
-    client: &Client<N, R, S>,
-    note_id: &str,
-    filename: Option<PathBuf>,
-) -> Result<File, String> {
-    let note_id = Digest::try_from(note_id)
-        .map_err(|err| format!("Failed to parse input note id: {}", err))?
-        .into();
-    let output_note = client
-        .get_output_notes(miden_client::store::NoteFilter::Unique(note_id))?
-        .pop()
-        .expect("should have an output note");
-
-    // Convert output note into input note before exporting
-    let input_note: InputNoteRecord = output_note
-        .try_into()
-        .map_err(|_err| format!("Can't export note with ID {}", note_id.to_hex()))?;
-
-    let file_path = filename.unwrap_or_else(|| {
-        let mut dir = PathBuf::new();
-        dir.push(note_id.inner().to_string());
-        dir
-    });
-
-    let mut file = File::create(file_path).map_err(|err| err.to_string())?;
-
-    file.write_all(&input_note.to_bytes()).map_err(|err| err.to_string())?;
-
-    Ok(file)
-}
-
-// IMPORT NOTE
-// ================================================================================================
-pub async fn import_note<N: NodeRpcClient, R: FeltRng, S: Store>(
-    client: &mut Client<N, R, S>,
-    filename: PathBuf,
-    verify: bool,
-) -> Result<NoteId, String> {
-    let mut contents = vec![];
-    let mut _file = File::open(filename)
-        .and_then(|mut f| f.read_to_end(&mut contents))
-        .map_err(|err| err.to_string());
-
-    // TODO: When importing a RecordedNote we want to make sure that the note actually exists in the chain (RPC call)
-    // and start monitoring its nullifiers (ie, update the list of relevant tags in the state sync table)
-    let input_note_record =
-        InputNoteRecord::read_from_bytes(&contents).map_err(|err| err.to_string())?;
-
-    let note_id = input_note_record.id();
-    client
-        .import_input_note(input_note_record, verify)
-        .await
-        .map_err(|err| err.to_string())?;
-
-    Ok(note_id)
 }
 
 // SHOW NOTE
@@ -498,10 +406,7 @@ mod tests {
     };
     use uuid::Uuid;
 
-    use crate::cli::{
-        get_input_note_with_id_prefix,
-        notes::{export_note, import_note},
-    };
+    use crate::cli::{export::export_note, get_input_note_with_id_prefix, import::import_note};
 
     #[tokio::test]
     async fn test_import_note_validation() {