+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 * Add `consumer_account_id` to `InputNoteRecord` with an implementation for sqlite store.
+* Added more feedback information to commands `account new`, `input-notes import`, `tx new` and `sync`.
 * Renamed the cli `input-notes` command to `notes`. Now we only export notes that were created on this client as the result of a transaction.
 * Added validation using the `NoteScreener` to see if a block has relevant notes.
 * Added flags to `init` command for non-interactive environments